refactor: move the upgrade-readiness service layer into internal/upgrade
The feature was smeared across three layers: pure checks in
pkg/upgradereadiness, the runner in internal/audit, and collectors + seam +
memo in internal/server. internal/upgrade now owns the whole internal service
layer — runner (upgrade.RunFromCache; posture audit and upgrade readiness are
deliberately distinct engines and no longer share a package), live collectors,
the EvidenceAuthorizer seam, ResolveHelmNamespaces, and the bounded scan memo
(upgrade.ScanMemoized) — mirroring how internal/capacity owns Karpenter.

internal/server keeps only the HTTP pieces (handler, response envelope, error
mapping, httpUpgradeAuthorizer over Server helpers); internal/mcp now depends
on internal/upgrade instead of the HTTP layer. audit exports
CollectTypedInput/ListNamespaced for the shared typed-input collection, and
allNamespaceNames moves to k8s.AllNamespaceNames for both consumers. Import
graph stays acyclic: server → upgrade → audit, mcp → upgrade.

No behavior change: full test suite green, and a live GKE scan post-move
matches the pre-move scan's verdict, summary, and per-check statuses and
finding counts.
