+++ internal/upgrade/collectors_test.go
@@ -1,4 +1,4 @@
-package server
+package upgrade
 
 import (
 	"context"
